Record Metrics for S3 Uploads (#1149)
This PR changes the instance cache to record metrics instead of issuing
error alerts when uploading to S3 fails. This is because uploading can
intermittently fail, and these instance are used for debugging anyway,
so temporary failures to upload is not a big issue. With metrics, we can
setup an alert if it starts failing for, say, 10 minutes straight which
means that there is probably something more seriously wrong that
requires some attention.

### Test Plan

CI