Prioritize updates of market order native token prices (#1162)
The native price cache cannot keep up with the amount of tokens it has
to process in production. This can lead to market orders not settling
because the orders are not considered solvable because of missing native
prices.

This PR prioritizes the updates of market order tokens over other
tokens. This functionality is only used by the autopilot's solvable
order module.