Make sure geth is killed in driver test (#1126)
This is a follow up to a discussion in a previous PR. I carefully
checked the tokio docs and code. Tokio does not wait for spawned tasks
to complete when the runtime is dropped (when a tokio test returns). In
order to ensure that the geth process really is killed, we have to wait
for it.

The most correct way to implement this would be to not use Drop, but
make tests manually kill geth. That's error prone or requires more test
setup (like a helper function that does it for you) so I went with
blocking in Drop.

Also changed the order of process killing and directory deletion because
it ensures the geth process won't try to use the directory while it is
being deleted.