Codebase-wide formatting (#1167)
Update the formatting on the whole codebase to match the driver, like
discussed before. This does not change default options, but rather
standardizes formatting for things that are otherwise not formatted
consistently.

This will cause merge conflicts. Once this is approved, I'm going to
ping everybody with the exact date when I intend to merge this so we can
minimize the impact.