Refactor uniswapv2-like liquidity
This changes the way we create uniswapv2-like liquidity to not use macros and modules. The motivation is that I wanted to add a custom univ2 liquidity argument, like what we have in the new solvers crate, to other crates. This turned out surprisingly tedious so I made this refactor first.