Remove MIP solver (#1179)
We already disabled the MIP solver and its configurations have been
removed. This PR removes all code related to the MIP solver.

I merged Quasimodo's `uses_internal_buffers` parameter with the external
solvers' one (we were using MIP's for that). This is because right now
we are configuring the two parameters to be the same and Quasimodo has
become more of an external solver anyway.

We also have an option to change the `time_limit` in the request that we
send to our solvers from integer to float. I didn't do it because other
external solvers right now might be relying on the fact that this is an
integer. Maybe it makes sense to discuss with the solver team if other
solvers would welcome this change?

*Note:* PR requires [these configuration
changes](https://gitlab.gnosisdev.com/devops/aws/aws-infra-staging/k8s-team-namespaces/gnosis-staging-k8s-gp/-/merge_requests/682)
before merging.

### Test Plan

Existing tests.