Get gas price in driver (#1214)
Fixes #1196 .

The driver was not fetching gas prices but they are required for the
solvers.

- Add some logging to the `route` function. This is useful because it
doesn't make sense to report on all errors in detail through the
autopilot-driver API but we do want to log them for debugging.
- Change the score comparison in the test to be approximate. Required
because the computed score is not deterministic (unrelated to my
change). The test fails if the score deviates by more than 1% from the
expected value.