Check Chain ID on Startup (#1219)
A couple of times, we have accidentally connected services meant for
Goerli to Mainnet. This leads to a bunch of events getting indexed in
the wrong database and is very painful to cleanup.

This PR adds a chain ID sanity check on start up to prevent this from
happening.

### Test Plan

```
% cargo run -p autopilot -- --node-url https://goerli.infura.io/v3/$INFURA_PROJECT_ID --chain-id 6
...
2023-02-15T15:40:45.363Z ERROR shared::tracing: thread 'main' panicked at 'assertion failed: `(left == right)`
  left: `5`,
 right: `6`: connected to node with incorrect chain ID'
```

### Release Notes

We should configure the `CHAIN_ID` environment variable on all our pods.