Remove order maturity code (#1221)
We decided to no longer use this concept and the maturity age is set to
0 in our cluster. I don't think we're going to bring it back so we
should remove the code. Correct me if I am wrong about it not coming
back.

Note that the is_mature field stays in the http solver api for now
because not including it would be a breaking change.