Custom 'solve' URL Path for HTTP Solvers (#1224)
I did something bad...

This **incredibly hacky** PR changes the Yearn solver price estimator to
use a custom path for the solve request. This is because the HTTP solver
expects requests to `/quote` instead of `/solve` for price estimation.

Why is it done in such a hacky way? Well the new `driver` binary already
supports arbitrary paths for solvers (so in this case we would be able
to configure the URL to be
`https://the.yearn.solver.eth/path/to/quote`). This means that this
technical debt will be removed for free in the very near future without
any additional work.

The alternative would be to stop appending `/solve` in the
`http_solver.rs` implementation and to add `/solve` to the URL for each
external solver in our deployment configuration. I chose against this
for a couple of reasons:
1. It is a configuration hassle with a non-0 risk.
2. External solvers have a `/notify` URL that gets added to the base for
solver competition. This has some pretty ambiguous semantics when paired
with changing the `DefaultHttpSolverApi` to be configured with a "solve"
URL instead of a base URL: is `/notify` appended to the solve URL, does
it `s/solve/notify/`? etc.

As mentioned ☝️, since this is already solved in the new driver, in a
correct and non-hacky way, I decided to use this hacky patch for now.

### Test Plan

Run the orderbook API and see it make requests with different paths:
```
% cargo run -p orderbook -- --price-estimators Yearn,Quasimodo --native-price-estimators Yearn,Quasimodo --yearn-solver-url http://localhost:10001 --yearn-solver-path quote --quasimodo-solver-url http://localhost:10002
...
2023-02-16T16:08:14.652Z DEBUG request{id=0}: shared::price_estimation::competition: new price estimate query=Query { from: None, sell_token: 0xdac17f958d2ee523a2206206994597c13d831ec7, buy_token: 0x6b175474e89094c44da98b954eedeac495271d0f, in_amount: 1000000000, kind: Sell } result=Err(Other(failed to send request

Caused by:
    0: error sending request for url (http://localhost:10001/quote?instance_name=2023-02-16_16%3A08%3A14.644312_UTC_Ethereum___Mainnet_1_0&time_limit=2&max_nr_exec_orders=100&use_internal_buffers=false&objective=surplusfeescosts): error trying to connect: tcp connect error: Connection refused (os error 61)
    1: error trying to connect: tcp connect error: Connection refused (os error 61)
    2: tcp connect error: Connection refused (os error 61)
    3: Connection refused (os error 61))) estimator="Yearn"
2023-02-16T16:08:14.652Z DEBUG request{id=0}: shared::price_estimation::competition: new price estimate query=Query { from: None, sell_token: 0xdac17f958d2ee523a2206206994597c13d831ec7, buy_token: 0x6b175474e89094c44da98b954eedeac495271d0f, in_amount: 1000000000, kind: Sell } result=Err(Other(failed to send request

Caused by:
    0: error sending request for url (http://localhost:10002/solve?instance_name=2023-02-16_16%3A08%3A14.646544_UTC_Ethereum___Mainnet_1_0&time_limit=2&max_nr_exec_orders=100&use_internal_buffers=false&objective=surplusfeescosts): error trying to connect: tcp connect error: Connection refused (os error 61)
    1: error trying to connect: tcp connect error: Connection refused (os error 61)
    2: tcp connect error: Connection refused (os error 61)
    3: Connection refused (os error 61))) estimator="Quasimodo"
...
```

**Note**: The errors are expected because I don't have anything actually
bound to those ports, but the paths are different for both price
estimators as expected.

### Release Notes

We should set `YEARN_SOLVER_PATH` to `quote` once it they switch their
routing.