Instrument Tenderly API (#1245)
Fixes #1241.

This PR adds a new instrumented Tenderly API implementation.
Additionally, we attach a "name" to each instance so that we have a
separate Gauge measuring which component is making which Tenderly
simulation.

Note that I also removed `TenderlyApi::block_number`. It was only used
in one place (and wasn't necessary - not specifying a block for a
tenderly simulation is the same as specifying the latest block, which is
what it was doing). I included this change here to not require
instrumenting that call as well.

### Test Plan

Run one of the binaries and check that we get metrics:
```
% curl -s 'http://localhost:9587/metrics' | rg tenderly
# HELP gp_v2_solver_tenderly_simulations Tenderly simulations.
# TYPE gp_v2_solver_tenderly_simulations gauge
gp_v2_solver_tenderly_simulations{name="access_lists",result="ok"} 1
```