Refactor e2e tests to use crate run functions (#1229)
This PR makes all e2e tests use the library run functions instead of
manually creating components. This separates the test from the code more
which will help future refactoring and makes the tests more extensive
because the components are run just like they are as standalone
binaries.

As a side effect many tests become simpler. There is potential for a lot
more e2e improvements after this but I didn't want to put more than
necessary in this PR.

E2e tests become slower because we can no longer drive individual
components at the exact moment it is needed. Instead some checks have to
be converted to polling with the `wait_for_condition` function. Running
all tests on CI goes from 90s to 120s. This is not a big deal. There are
low hanging fruits to speed tests up and eventually the plan is to run
e2e tests in parallel.

Other changes in this PR that I did not factor out:
- Improve logging. I was looking at a lot of logs while debugging the
tests and in the process moved some to trace when they were large and
printing duplicate information. We can always change some of these back
later if they turn out to have been useful. Other logs were missing
information.
- Make balancer contract address configurable. Needed for e2e test.