Cache cow reference price for rewards calculation (#1257)
Fixes #1238.

Describe context of what this change does, why it is needed and how it
is accomplished
This change introduces a fallback cache for the CoW token price inside
one of autopilot's components. The CoW price is used to calculate the
risk adjusted solver rewards
([CIP-14](https://forum.cow.fi/t/cip-14-risk-adjusted-solver-rewards/1132))
when building an auction. Therefore, if the autopilot fails to estimate
this price, it cannot update the current auction and the protocol stops.
This cache will mitigate transient errors (e.g. network failures when
fetching CoW price from other APIs) at the cost of using a stale CoW
price.

### Test Plan

A unit test for the `BestEffortCowPriceEstimator` struct. Existing
tests.