SolverRewards 9: Update settlement details if exist (#1294)
All settlement events are indexed by both staging and production
databases.
However, auctions only exist in one of the environments.

This PR inserts settlement observations only if the auction exist.
Thanks @MartinquaXD for figuring out the issue. I did have the same
issue when ran CIP20 changes locally, but thought it is because of
running the local db. :/