Improve bool argument handling (#1300)
By default a clap bool argument is disabled and enabled when the flag is
specified like `--foo`. This makes it impossible to define a flag that
is default enabled while also being able to be manually turned off with
something like `--foo=false`. In this PR I change all bool arguments to
require passing in a value. This is more consistent with how other
arguments are parsed and more flexibly because it allows default
enabled. This changes that that a plain `--foo` without value is not
longer accepted, which is fine.