`driver`: Verify Solver Account Balance (#1365)
This PR adds an additional check for solver balances before committing
to a solution.

Previously, we had relied on simulations from the node failing on
`insufficient funds for gas * price + value` errors. This was
problematic for two reasons:
* We weren't specifying gas parameters correctly when creating access
lists. This lead to the default `gas` and `gas_price` values to be
chosen by the node and Geth seems to use unreasonable defaults (cc
@mateipopa - this solves the issue that we saw on Görli connected to
NodeReal, where we were not submitting settlements because of
insufficient balances despite having a good amount of Ether and network
conditions being "normal")
* These errors are node-dependent and not guaranteed by the Ethereum RPC
standard. Specifically, Nethermind does not check account gas values on
simulations (see
https://github.com/cowprotocol/services/pull/1361#issuecomment-1477951819).

### Test Plan

Added an E2E test that verifies that the `driver` responds with a
correct error code when the solver account has insufficient balance for
the proposed settlement.

---------

Co-authored-by: Martin Beckmann <martin.beckmann@protonmail.com>