Fix error when single order solver fee is larger than sell amount (#1418)
Fixes https://github.com/cowprotocol/services/issues/1417

This code has a special case for partially fillable sell orders in the
match statement at line 333 ` let (executed_sell_amount,
executed_buy_amount) = match order.kind { ...`. For buy orders the sell
amount is simply increased by the fee. For sell orders if we did the
same thing, then we could end up trying to sell more than order's sell
amount, which is impossible.

For this reason the resulting sell amount is capped with `
.min(order.sell_amount)`. When we hit the cap, we can't increase the
sell amount more, so we decrease the buy amount so that less of the
order is sold while keeping the price the same.

This part of the code errors when it can't subtract surplus_fee from
executed_sell_amount. Instead of erroring, it should return `Ok(None)`
because it isn't an error. It just means that there isn't enough space
in the sell amount to account for the fee.